turn is now being updated on client and firestore

+++ src/main/java/Models/Turn.java
@@ -60,7 +60,7 @@ public class Turn implements Model, FirestoreFormattable, Observer<DocumentSnaps
             assert map != null;
             activePlayer = Players.getByStringUuid((String) map.get("activePlayer"))
                     .orElseThrow(() -> new PlayerException("Player Id doesn't exist."));
-            amountOfDouble = (int) map.get("amountOfDoubles"); //Shouldn't this cast to (long) instead of (int)? -Vincent
+            amountOfDouble = (int) (long) map.get("amountOfDoubles"); //Shouldn't this cast to (long) instead of (int)? -Vincent
         } catch(PlayerException playerException) {
             playerException.printStackTrace();
         }
